+++ CHANGELOG.md
@@ -11,9 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [research-pillars] Moved GetSubjectDataRaw to GetSubjectData
+
 ### Deprecated
 
 ### Removed
+
 
 ### Fixed